Utils  * Added more `MutableXXX` classes.
 * Fixed the assertions of `CollectionUtils.sort(...)`.

 * Revamped `CSVLine` to use records.
 * `CSVReader` now fill the lines with `null` values so that striped values and missing columns can be distinguished.

 * Renamed the parser of interger numbers (`TextParsers.integer(...)`).
 * Added a parser of decimal numbers (`TextParsers.decimal(...)`).

+++ work/Tests/module/sources/com/trazere/test/ParserTest.java
@@ -195,7 +195,7 @@ public class ParserTest {
 	};
 	static {
 		final List<Parser<Character, ? extends Integer>> exprs = new ArrayList<Parser<Character, ? extends Integer>>();
-		exprs.add(TextParsers.decimal());
+		exprs.add(TextParsers.integer());
 		exprs.add(EXPR_PAREN);
 		exprs.add(EXPR_PLUS);
 		exprs.add(EXPR_MUL);
